You can now find how many passes occur between two players, by calling the getQueryNumOfPassesBetweenPlayers() function in the databaseController class.
This is demonstrated in the main method.

This is unfortuantely for players across all 6 games and not just a single game, this is the next thing I will fix.

Additionally, I will add getting all the locations of every pass between two players during a game.

Now I am tired and going to bed, don't look at the time of this commit.